Added a cvar to enable recording mvd files containing 'unsafe' extensions, so CSQC ents can now be recorded. This defaults to off, and is expected to be enabled for mods that are truely FTE only. Recording qwd files works again. Not all extensions are fully implemented, when recording both mvds and qwds, particularly the initial connection info (including model list length). Ezquake's invalid chunk requests generate no protocol response. Ezquake should no longer kick itself after every download. Couple of other bugs fixed.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@3893 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_parse.c
@@ -3200,7 +3200,7 @@ void CL_ParseBaseline2 (void)
 {
 	entity_state_t es;
 
-	CL_ParseDelta(&nullentitystate, &es, MSG_ReadShort(), true);
+	CL_ParseDelta(&nullentitystate, &es, (unsigned short)MSG_ReadShort(), true);
 	if (!CL_CheckBaselines(es.number))
 		Host_EndGame("CL_ParseBaseline2: check baselines failed with size %i", es.number);
 	memcpy(cl_baselines + es.number, &es, sizeof(es));
